requirements.txt
rasterio
pytorch
tqdm
arcpy
dotenv